// src/interfaces/eth-networks.ts
export type TransactionResut = {
  blockHash: string;
  blockNumber: string;
  chainId: string;
  condition: string;
  creates: string;
  from: string;
  gas: string;
  gasPrice: string;
  hash: string;
  input: string;
  nonce: string;
  publicKey: string;
  r: string;
  raw: string;
  s: string;
  standardV: string;
  to: string;
  transactionIndex: string;
  v: string;
  value: string;
};

export type ETHTransactionStatus = {
  status: string;
};